V1/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
statsmodels==0.14.0
ta==0.10.2
sqlite3
requests==2.31.0
groq==0.4.1
